move string functions to helper file
