+++ clang_plugin/src/helpers.h
@@ -1,0 +1,5 @@
+#include <clang-c/Index.h>
+
+char *concat_strings(const char *s1, const char *s2);
+
+const char *unwrap_string(CXString cxStr);
